Feat: Refactoring old ecommerce api
